Moved AdvectField plugin into a new module for generating nowcast forecasts.
