+++ bin/improver-nowcast-extrapolate
@@ -38,7 +38,7 @@ import numpy as np
 
 from iris import Constraint
 from improver.argparser import ArgParser
-from improver.nowcasting.optical_flow import AdvectField
+from improver.nowcasting.forecasting import AdvectField
 from improver.nowcasting.utilities import ApplyOrographicEnhancement
 from improver.utilities.filename import generate_file_name
 from improver.utilities.load import load_cube
