+++ lib/improver/tests/nowcasting/forecasting/test_AdvectField.py
@@ -39,7 +39,7 @@ from iris.coords import DimCoord
 from iris.exceptions import InvalidCubeError
 from iris.tests import IrisTest
 
-from improver.nowcasting.optical_flow import AdvectField
+from improver.nowcasting.forecasting import AdvectField
 from improver.tests.set_up_test_cubes import set_up_variable_cube
 from improver.utilities.warnings_handler import ManageWarnings
