+++ lib/improver/tests/nowcasting/optical_flow/test_check_input_coords.py
@@ -37,7 +37,7 @@ from iris.coords import DimCoord
 from iris.exceptions import InvalidCubeError
 from iris.tests import IrisTest
 
-from improver.tests.nowcasting.optical_flow.test_AdvectField \
+from improver.tests.nowcasting.forecasting.test_AdvectField \
     import set_up_xy_velocity_cube
 from improver.nowcasting.optical_flow import check_input_coords
